+++ lib/src/toast/lib.dart
@@ -1,1 +1,2 @@
+export 'scrim_host.dart';
 export 'toast.dart';